+++ rsa/pkcs1.py
@@ -262,7 +262,12 @@ def decrypt(crypto: bytes, priv_key: key.PrivateKey) -> bytes:
         sep_idx = cleartext.index(b'\x00', 2)
     except ValueError:
         sep_idx = -1
-    sep_idx_bad = sep_idx < 0
+    # sep_idx indicates the position of the `\x00` separator that separates the
+    # padding from the actual message. The padding should be at least 8 bytes
+    # long (see https://tools.ietf.org/html/rfc8017#section-7.2.2 step 3), which
+    # means the separator should be at least at index 10 (because of the
+    # `\x00\x02` marker that preceeds it).
+    sep_idx_bad = sep_idx < 10
 
     anything_bad = crypto_len_bad | cleartext_marker_bad | sep_idx_bad
     if anything_bad: